requirements.txt
asttokens==3.0.1
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
ipython==8.18.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.6
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyYAML==6.0.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
